Work around navbar menu CSS cascade bug
Our frontend build has a [bug](https://trello.com/c/VGvOyHk4) where CSS
modules are included multiple times in bundles when they are imported
both from JavaScript/Elm and CSS/Sass.

As a result of this bug, a second copy of the Tooltip CSS module's styles is being
output after the Menu CSS module, which depends on it. This causes
selectors in Menu that should take precedence due to source order to be
overridden by selectors of the same specificity in Tooltip.

This hack works around one such bug until we are able to correct the CSS
output duplication.