Make sass-loader a peerDependency
We have experienced intermittent hangs in the final stages of webpack
compilation when two versions of sass-loader are present in the build.
By making sass-loader a peerDependency, we hope to ensure that the
consuming project will specify a compatible version range, enabling
Yarn/NPM to select a single version of sass-loader.