Use NotificationState type rather than hidden Bool
This allows the same amount of configuration at the callsite (store a single value, and respond to a single event, but now a NotificationState type rather than a Bool).

But now we can correctly animate our states and correctly remove from the DOM when the state is "Hidden".

This change gets Visible, Disappearing and Hidden states correctly working.

The "Appearing" state is currently not working because I haven't figured out how to do the equivalent of `requestAnimationFrame()`, and I'm on a flight so googling is hard.