Add proper support for "Appearing" state, requiring a large refactor
The "Appearing" state works by having a subscription that, if any notifications are in the "Appearing" state, will transition them to "Visible" on the next animation frame.

To achieve this the subscription reads the model and looks for existing notifications states, so we needed to ensure that all notification states were known when during the `init` phase.

To achieve this, I updated the `JsxDecoder` code to have a more complex `JsxWithMessageDecoder`, which can emit messages during an initial pass of the JSX tree, allowing us to initiate the model correctly.

- Added `Notification.subscriptions` so we can subscribe to the `Appearing -> Visible` transitions.
- Added a `Notification.getAutomationId` helper, mainly for use in the Notification Demos
- Updated Notification.Demo and the Notification Json Decoder to correctly initialise state, store it in a dictionary, and update it as needed.
- Updated Demo helpers to have more complex `jsxChildrenWithMessages` to allow initialising state on the first pass.
- Refactored Demo helpers to have clearer types (type aliases, better use of "let" definitions)
- Updated presets to have automationIds, so our Notification Demo app can correctly manage their state
- Add "animation-frame" dependency