Get autohide working for Toast Notifications
At first I thought this wouldn't be possible, because when we set the "Disappearing" state on a timer, we didn't have access to a MouseEvent from which we could determine the height of the animation in order to animate it. I worked around this by using an "transitionstart" event.

I also tossed and turned on whether to change our Notification.subscription function so that it had access to our Notification Config in order to know whether it was autohide. Instead, I chose to use a new `Autohide` notification state.

This came with the following trade-offs:

- Positive: when consuming the Notification, if you want subscriptions you only need to provide the current state and a setter to update the state. If we needed to provide the config, you would need to update your application state model to include the Config for all notifications, which would have been quite messy.

- Negative: this allows you to create a Global or Inline notification with the state "Autohide Appearing", which is a state I would prefer to be impossible.