Separate NotificationState and NotificationStage
Previously NotificationState was a recursive union type, with `Autohide` being a special state that contained the other states. This allowed invalid types though, like `Autohide(Autohide(Autohide(Visible)))`.

Now instead we have NotificationStage, and NotificationState.

I've also made it so NotificationState is passed in with the view function, rather than being something optionally added to the config.

I think the mechanism for updating state will need to be in there too - but I'll do that as a separate PR.