fix(webpack): No longer include focus-visible polyfill by default
The decorator added it to all entrypoints, including entrypoints which are other polyfills - which is not desirable. Now we have put the expectation back on the consuming project (murmur) to add the polyfill themselves.

If it's not present, the style-guide components still work but without the additional accessibility / keyboard focus styles.