Fix elmSvgAssetLoader to not rely on elm being installed in the style guide.
Previously this was checking the file `node_modules/elm/package.json` to get the
elm version, but this only existed because the elm-webpack-loader was a dependency.
Now that it's not, that file doesn't exist, and so the SVG-loader causes build errors.

Instead of trying to autodetect the Elm version, I'm going to assume Elm 0.18 unless
an `isElm19` config flag is specified. As we upgrade murmur apps to Elm 19 they won't be
using this loader anyway, they'll be using the new Babel plugin instead, so that's fine.